Break HTTP callback out into handlers
This breaks http callbacks into separate handlers for each major API
node. In doing so, it adds error handling for scenes so that clients
recover faster versus an empty response and adds support for group 0.
This also adds more robust URI dissassembly and fakes support for
new light searches. With this patch in place, the new Hue app is
able to interact at least minimally with the emulated bridge, though
it cannot yet control lights because it needs lights organized into
rooms (groups).