Refactor shared code and add missing response (#66)
The light state put/post was missing a response entirely on success
which was causing some 3rd part apps to be angry. According to the API,
"Updated." is a valid reponse for this method, but is not accepted by
some apps. This adds the full response for light state put/post and
consolidatess response generation code.