Fix several bugs (#74)
* LightService: Remove unnecessary String object

This object is already a string, no need to dup it into a new string
structure.

* SSDP: Use correct UUID suffix

Using the MAC address as the UUID suffix prevents multiple Hue devices
from showing up with the same IP address when scanning with 3rd party
apps.

* SSDP: Rework the search and notify parsing

This significantly simplifies and modularizes the M-SEARCH and NOTIFY
parsing code while fixing a couple of bugs in parsing. This code is now
readable.